Tech/80641 attachment component (#511)
* Allow uploading multiple files at once in AttachmentList component

* Added drag and drop field to attachment component

* Tests for AttachmentList

* Cleanup

* Added another test to-do and added size prop to attachment list

* Start of getting correct headers and columns in AttachmentList

* started replacing attachments in createAndEditIPO with AttachmentList component

* The two places AttachmentList was used mostly work with the updated component

* finished replacing attachments in CreateAndEditIPO

* Removed column type in AttachmentList

* AttachmentList component finished?

* Replaced attachment list in viewIPO with AttachmentList component

* fix column widths in viewIPO attachments

* Add return type to getColumns function

* tests

* Cleanup

* Fixed failing test

* Started AttachmentList storybook

* Added storybook stories for AttachmentList component

* Fixed AttachmentList stories

+++ src/modules/Preservation/views/ScopeOverview/TagFlyout/ActionTab/ActionAttachments.tsx
@@ -1,10 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import { showSnackbarNotification } from '../../../../../../core/services/NotificationService';
 import { usePreservationContext } from '../../../../context/PreservationContext';
-import AttachmentList, { Attachment } from '@procosys/components/AttachmentList';
+import AttachmentList from '@procosys/components/AttachmentList';
 import { Canceler } from 'axios';
 import Spinner from '@procosys/components/Spinner';
 import { Container } from './ActionAttachment.style';
+import { Attachment } from '@procosys/modules/InvitationForPunchOut/types';
+import { TableOptions } from 'react-table';
 
 interface ActionAttachmentsProps {
     tagId: number;
@@ -51,43 +53,58 @@ const ActionAttachments = ({
         };
     }, []);
 
-
-    const downloadAttachment = async (attachmentId: number): Promise<void> => {
-        try {
-            const url = await apiClient.getDownloadUrlForActionAttachment(tagId, actionId, attachmentId);
-            window.open(url, '_blank');
-            showSnackbarNotification('Attachment is downloaded.', 5000, true);
-        } catch (error) {
-            console.error('Not able to get download url for action attachment: ', error.message, error.data);
-            showSnackbarNotification(error.message, 5000, true);
+    const downloadAttachment = async (attachment: Attachment): Promise<void> => {
+        if(attachment.id){
+            try {
+                const url = await apiClient.getDownloadUrlForActionAttachment(tagId, actionId, attachment.id);
+                window.open(url, '_blank');
+                showSnackbarNotification('Attachment is downloaded.', 5000, true);
+            } catch (error) {
+                console.error('Not able to get download url for action attachment: ', error.message, error.data);
+                showSnackbarNotification(error.message, 5000, true);
+            }
+        }else{
+            console.error('Cannot download an attachment that does not exist');
+            showSnackbarNotification('Cannot download an attachment that does not exist', 5000, true);
         }
     };
 
-    const addAttachment = async (file: File): Promise<void> => {
-        try {
-            setIsLoading(true);
-            await apiClient.addAttachmentToAction(tagId, actionId, file, false);
-            getAttachments();
-            showSnackbarNotification(`Attachment with filename '${file.name}' is added to action.`, 5000, true);
-        } catch (error) {
-            console.error('Upload file attachment failed: ', error.message, error.data);
-            showSnackbarNotification(error.message, 5000, true);
+    const addAttachments = async (files: FileList | null): Promise<void> => {
+        if (!files) {
+            showSnackbarNotification('No files to upload');
+            return;
         }
+        setIsLoading(true);
+        Array.from(files).forEach(async file => {
+            try {
+                await apiClient.addAttachmentToAction(tagId, actionId, file, false);
+                getAttachments();
+                showSnackbarNotification(`Attachment with filename '${file.name}' is added to action.`, 5000, true);
+            } catch (error) {
+                console.error('Upload file attachment failed: ', error.message, error.data);
+                showSnackbarNotification(error.message, 5000, true);
+            }
+        });
         setIsLoading(false);
     };
 
-
-    const deleteAttachment = async (attachment: Attachment): Promise<void> => {
-        try {
-            setIsLoading(true);
-            await apiClient.deleteAttachmentOnAction(tagId, actionId, attachment.id, attachment.rowVersion);
-            getAttachments();
-            showSnackbarNotification(`Attachment with filename '${attachment.fileName}' is deleted.`, 5000, true);
-        } catch (error) {
-            console.error('Not able to delete action attachment: ', error.message, error.data);
-            showSnackbarNotification(error.message, 5000, true);
-        }
-        setIsLoading(false);
+    const deleteAttachment = async (row: TableOptions<Attachment>): Promise<void> => {
+        const attachment: Attachment = row.value;
+        if(attachment.id && attachment.rowVersion){
+            try {
+                setIsLoading(true);
+                await apiClient.deleteAttachmentOnAction(tagId, actionId, attachment.id, attachment.rowVersion);
+                getAttachments();
+                showSnackbarNotification(`Attachment with filename '${attachment.fileName}' is deleted.`, 5000, true);
+            } catch (error) {
+                console.error('Not able to delete action attachment: ', error.message, error.data);
+                showSnackbarNotification(error.message, 5000, true);
+            }
+            setIsLoading(false);
+        }else{
+            console.error('Cannot delete an attachment that does not exist');
+            showSnackbarNotification('Cannot delete an attachment that does not exist', 5000, true);
+        };
     };
 
     if (isLoading) {
@@ -101,7 +118,7 @@ const ActionAttachments = ({
             <AttachmentList
                 attachments={attachments}
                 disabled={isVoided}
-                addAttachment={enableActions ? addAttachment : undefined}
+                addAttachments={enableActions ? addAttachments : undefined}
                 deleteAttachment={enableActions ? deleteAttachment : undefined}
                 downloadAttachment={downloadAttachment}
             />
